Use UTF-8 as the charset for HttpTemplate when decoding the response body String.
When decoding the response, the default platform charset was being used. Relying on runtime configuration is discouraged and an explicit value is preferred. An explicit UTF-8 value is already being used in this same file, so this was likely an oversight. It is expected that UTF-8 is already the platform default for most deployments.